Fix full deque iteration bug
Prior to this change when the deque was full the begin and end iterators would immediately be equal, thus preventing iteration. This change adds a counter to each iterator that tracks the number of items that have been iterated through and uses that counter as part of the comparison, thus an end iterator's counter will match the size of the deque.